move 'test_repartitioning_after_concat' test
